3.2.Sante-diagnostique/requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
Pillow>=10.0.0
scikit-learn>=1.3.0

# PyTorch
torch>=2.2.0
torchvision>=0.17.0

# Notebook execution
nbclient>=0.7.4
nbformat>=5.9.0

# Front-end
streamlit>=1.22.0

# Image processing
opencv-python>=4.7.0

# Utilities
pathlib>=1.0.1
tqdm